Move RFileUtil to datawave-core and use in SplittableRFileRecordReader
